Argument testing on more functions...
Split out test_intfns.py into its module components
